AzureServerlessTemplate/CloudRetrievalViaDask/lambda/requirements.txt
azure-functions
azure-mgmt-compute
azure-mgmt-resource
azure-identity
azure-cli-core